+++ src/common-capabilities/generate-route-guard.ts
@@ -1,21 +1,23 @@
-import { promptInput } from '@extensionFramework';
-import { isKebabCase, kebabCaseToCamelCase, kebabCaseToPascal } from '@utils';
-import { generateAngularElement, generateSpec } from './util';
+import { toCamelCase, toPascalCase, toKebabCase } from '@utils';
+import { generateAngularElement, generateSpec, promptForName } from './util';
 import {
 	RouteGuardSpecTemplateData,
 	RouteGuardTemplateData,
 	TemplateFileNames,
 } from '@models';
 import * as path from 'path';
+import { AngularElement } from './models';
 
 /**
  * @param folderRightClickedPath /home/fernando/test/src/app
  */
 export const generateRouteGuard = async (folderRightClickedPath: string) => {
-	const nameInKebabCase = await promptForName();
+	const nameInKebabCase = toKebabCase(
+		await promptForName(AngularElement.GUARD, 'user-auth'),
+	);
 	const templateData: RouteGuardTemplateData = {
-		className: `${kebabCaseToPascal(nameInKebabCase)}Guard`,
-		fnName: `${kebabCaseToCamelCase(nameInKebabCase)}Guard`,
+		className: `${toPascalCase(nameInKebabCase)}Guard`,
+		fnName: `${toCamelCase(nameInKebabCase)}Guard`,
 	};
 
 	await generateAngularElement(
@@ -26,24 +28,13 @@ export const generateRouteGuard = async (folderRightClickedPath: string) => {
 	);
 };
 
-const promptForName = async (): Promise<string> => {
-	return await promptInput({
-		prompt: 'Enter route guard name (kebab-case)',
-		placeHolder: 'e.g. user-auth',
-		validationFn: value =>
-			isKebabCase(value)
-				? null
-				: 'Route guard name must be in kebab-case format',
-	});
-};
-
 export const generateRouteGuardSpec = async (
 	routeGuardFilePath: string,
 ): Promise<void> => {
 	const nameInKebabCase = path.basename(routeGuardFilePath, '.guard.ts');
 	const templateData: RouteGuardSpecTemplateData = {
-		className: `${kebabCaseToPascal(nameInKebabCase)}Guard`,
-		fnName: `${kebabCaseToCamelCase(nameInKebabCase)}Guard`,
+		className: `${toPascalCase(nameInKebabCase)}Guard`,
+		fnName: `${toCamelCase(nameInKebabCase)}Guard`,
 		nameInKebabCase,
 	};