+++ src/common-capabilities/models/angular-element.enum.ts
@@ -1,0 +1,9 @@
+export enum AngularElement {
+	COMPONENT = 'component',
+	PIPE = 'pipe',
+	DIRECTIVE = 'directive',
+	SERVICE = 'service',
+	GUARD = 'guard',
+	INTERCEPTOR = 'interceptor',
+	RESOLVER = 'resolver',
+}