+++ src/test/suite/fixtures/index.ts
@@ -2,3 +2,4 @@ export * from './component';
 export * from './pipe';
 export * from './directive';
 export * from './service';
+export * from './route-guard';