+++ src/test/suite/generate-component.test.ts
@@ -52,7 +52,7 @@ suite('Generate Component', () => {
 		() => {
 			test('should generate a component and spec file', async () => {
 				createPromptStub(sandbox)
-					.inputBox('dummy') // "Enter component name (kebab-case)"
+					.inputBox('dummy') // "Enter component name"
 					.apply();
 
 				await runCommand();
@@ -78,6 +78,52 @@ suite('Generate Component', () => {
 				);
 			});
 
+			test('should support camel case component names', async () => {
+				createPromptStub(sandbox)
+					.inputBox('dummyInCamelCase') // "Enter component name"
+					.apply();
+
+				await runCommand();
+
+				const componentPath = path.join(
+					getSrcDirectoryPath(),
+					'dummy-in-camel-case.component.ts',
+				);
+				const specPath = path.join(
+					getSrcDirectoryPath(),
+					'dummy-in-camel-case.component.spec.ts',
+				);
+
+				assertItExists(
+					componentPath,
+					`Component file should exist at ${componentPath}`,
+				);
+				assertItExists(specPath, `Spec file should exist at ${specPath}`);
+			});
+
+			test('should support pascal case component names', async () => {
+				createPromptStub(sandbox)
+					.inputBox('DummyInCamelCase') // "Enter component name"
+					.apply();
+
+				await runCommand();
+
+				const componentPath = path.join(
+					getSrcDirectoryPath(),
+					'dummy-in-camel-case.component.ts',
+				);
+				const specPath = path.join(
+					getSrcDirectoryPath(),
+					'dummy-in-camel-case.component.spec.ts',
+				);
+
+				assertItExists(
+					componentPath,
+					`Component file should exist at ${componentPath}`,
+				);
+				assertItExists(specPath, `Spec file should exist at ${specPath}`);
+			});
+
 			test('should generate a component file, but using a custom template when the user provide it', async () => {
 				await makeAngularToolboxDirectory();
 				await createTemplateFile(
@@ -85,7 +131,7 @@ suite('Generate Component', () => {
 					customComponentTemplateTestingData,
 				);
 				createPromptStub(sandbox)
-					.inputBox('dummy') // "Enter component name (kebab-case)"
+					.inputBox('dummy') // "Enter component name"
 					.apply();
 
 				await runCommand();
@@ -108,7 +154,7 @@ suite('Generate Component', () => {
 					'showErrorMessage',
 				);
 				createPromptStub(sandbox)
-					.inputBox('dummy') // "Enter component name (kebab-case)"
+					.inputBox('dummy') // "Enter component name"
 					.apply();
 
 				await runCommand();
@@ -133,7 +179,7 @@ suite('Generate Component', () => {
 						skipSpec: true,
 					});
 					createPromptStub(sandbox)
-						.inputBox('dummy') // "Enter component name (kebab-case)"
+						.inputBox('dummy') // "Enter component name"
 						.apply();
 
 					await runCommand();
@@ -155,7 +201,7 @@ suite('Generate Component', () => {
 						},
 					});
 					createPromptStub(sandbox)
-						.inputBox('dummy') // "Enter component name (kebab-case)"
+						.inputBox('dummy') // "Enter component name"
 						.apply();
 
 					await runCommand();
@@ -183,7 +229,7 @@ suite('Generate Component', () => {
 						},
 					});
 					createPromptStub(sandbox)
-						.inputBox('dummy') // "Enter component name (kebab-case)"
+						.inputBox('dummy') // "Enter component name"
 						.apply();
 
 					await runCommand();
@@ -210,7 +256,7 @@ suite('Generate Component', () => {
 						},
 					});
 					createPromptStub(sandbox)
-						.inputBox('dummy') // "Enter component name (kebab-case)"
+						.inputBox('dummy') // "Enter component name"
 						.apply();
 
 					await runCommand();
@@ -228,7 +274,7 @@ suite('Generate Component', () => {
 					});
 					await setPrefixInWorkspaceConfig('prefix');
 					createPromptStub(sandbox)
-						.inputBox('dummy') // "Enter component name (kebab-case)"
+						.inputBox('dummy') // "Enter component name"
 						.apply();
 					const componentPath = path.join(
 						getSrcDirectoryPath(),
@@ -251,7 +297,7 @@ suite('Generate Component', () => {
 						skipPrefix: false,
 					});
 					createPromptStub(sandbox)
-						.inputBox('dummy') // "Enter component name (kebab-case)"
+						.inputBox('dummy') // "Enter component name"
 						.apply();
 
 					await runCommand();
@@ -272,7 +318,7 @@ suite('Generate Component', () => {
 						'showErrorMessage',
 					);
 					createPromptStub(sandbox)
-						.inputBox('dummy') // "Enter component name (kebab-case)"
+						.inputBox('dummy') // "Enter component name"
 						.apply();
 
 					await runCommand();
@@ -292,7 +338,7 @@ suite('Generate Component', () => {
 						},
 					});
 					createPromptStub(sandbox)
-						.inputBox('dummy') // "Enter component name (kebab-case)"
+						.inputBox('dummy') // "Enter component name"
 						.apply();
 
 					await runCommand();
@@ -319,7 +365,7 @@ suite('Generate Component', () => {
 						},
 					});
 					createPromptStub(sandbox)
-						.inputBox('dummy') // "Enter component name (kebab-case)"
+						.inputBox('dummy') // "Enter component name"
 						.apply();
 
 					await runCommand();